Add activity creator.
Sometimes is useful to have the concept of "activity feed" created by. The use
case for us is the following.

We have sites that have several users. The activity that occurs on a site is made by an user. So
the owner of the activity is the site but we want to track which user did that action on the site.