Set self.abstract_class = true
`base_class` in ActiveRecord returns
`PublicActivity::ORM::ActiveRecord::Activity` incorrectly. Since
`PublicActivity::ORM::ActiveRecord::Activity` is really an abstract
class for all intents and purposes, it should be declared so that
`base_class` returns `PublicActivity::Activity` instead.

My use case was using the `unread` gem to track who had read an
activity. It uses `base_class` to record the polymorphic type of the
`readable` which is the Activity in this case.

All the tests passed with this addition but I’ve added another to
ensure that `base_class` returns the correct value.