Mongoid - Activity relations optional support
With this change to optional in mongoid 7 and above support has been added by this commit https://github.com/public-activity/public_activity/commit/debdd74c6e82f55ab3f196419c53ee5c78f87b8a. But the test has optional as true but the actual code has optional as false. This is makinng upgrading to the latest version for this gem not feasible with newer mongoid version 7 and 8.